extract to fetcher class
