+++ src/main/kotlin/com/nisecoder/gradle/atcoder/internal/Entities.kt
@@ -1,0 +1,13 @@
+package com.nisecoder.gradle.atcoder.internal
+
+
+class ContestTaskList {
+    lateinit var tasks: List<ContestTask>
+}
+
+data class ContestTask(
+    val taskId: String,
+    val taskName: String,
+    val timeLimit: String,
+    val memoryLimit: String,
+)
